# src/main/java/com/utils/openapi/api/UserClientApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.utils.openapi.api;

import com.utils.openapi.model.UserRequestBody;
import com.utils.openapi.model.UserResponseBody;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-09-16T17:12:39.115251300+01:00[Africa/Lagos]")
@Validated
@Api(value = "UserClient", description = "the UserClient API")
public interface UserClientApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /api/v1/users/{userId}
     *
     * @param userId user id (required)
     * @return user returned successfully (status code 200)
     */
    @ApiOperation(value = "", nickname = "getUserById", notes = "", response = UserResponseBody.class, tags={ "user-client", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "user returned successfully", response = UserResponseBody.class) })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/users/{userId}",
        produces = { "application/json" }
    )
    default ResponseEntity<UserResponseBody> _getUserById(@ApiParam(value = "user id", required = true) @PathVariable("userId") String userId) {
        return getUserById(userId);
    }

    // Override this method
    default  ResponseEntity<UserResponseBody> getUserById(String userId) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"phoneNumber\" : \"phoneNumber\", \"id\" : \"id\", \"email\" : \"email\", \"username\" : \"username\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /api/v1/users
     *
     * @param userRequestBody  (required)
     * @return user updated successfully (status code 204)
     */
    @ApiOperation(value = "", nickname = "updateUser", notes = "", tags={ "user-client", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "user updated successfully") })
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/v1/users",
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> _updateUser(@ApiParam(value = "", required = true) @Valid @RequestBody UserRequestBody userRequestBody) {
        return updateUser(userRequestBody);
    }

    // Override this method
    default  ResponseEntity<Void> updateUser(UserRequestBody userRequestBody) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}